requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==25.0
requests==2.32.3
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
